src/backend/requirements.txt
Flask
Flask-SocketIO
eventlet
python-dotenv
requests
openai

